Created Data-Gen file
