+++ database/fakeData/fakeDataHelpers.js
@@ -1,6 +1,6 @@
-var faker = require('faker');
-var { database, user, password } = require('../config/config.js');
 const mysql = require('mysql');
+const Generate = require('./fakeDataGen.js');
+var { database, user, password } = require('../config/config.js');
 const connection = mysql.createConnection({
   host: 'localhost',
   user,
@@ -8,17 +8,8 @@ const connection = mysql.createConnection({
   database
 });
 
-
-const createFakeProductTypes = (n) => {
-  var productTypes = [];
-  for (var i = 0; i < n; i++) {
-    productTypes.push(faker.commerce.product());
-  }
-  return productTypes;
-};
-
 const fillProductTypes = async (n) => {
-  var productTypes = createFakeProductTypes(n);
+  var productTypes = Generate.createProductTypes(n);
 
   var qString = 'INSERT IGNORE INTO productTypes (Name) VALUES (?)';
   for (var i = 0; i < productTypes.length; i++) {
@@ -31,17 +22,8 @@ const fillProductTypes = async (n) => {
   // console.log('all stored');
 };
 
-//fill the categories table
-const createFakeCategories = (n) => {
-  var categories = [];
-  for (var i = 0; i < n; i++) {
-    categories.push(faker.commerce.department());
-  }
-  return categories;
-};
-
 const fillCategories = async (n) => {
-  var categories = createFakeCategories(n);
+  var categories = Generate.createCategories(n);
   var qString = 'INSERT IGNORE INTO categories (Name) VALUES (?)';
   for (var i = 0; i < categories.length; i++) {
     await connection.query(qString, [categories[i]], (err, res) => {
@@ -51,22 +33,8 @@ const fillCategories = async (n) => {
   }
 };
 
-const createFakeProducts = (n) => {
-  var products = [];
-  for (var i = 0; i < n; i++) {
-    products.push([
-      faker.commerce.productName(),
-      faker.commerce.price(),
-      faker.image.image(),
-      Math.floor(Math.random() * 2),
-      faker.commerce.department(),
-      faker.commerce.product()
-    ]);
-  }
-  return products;
-};
 const fillProducts = async (n) => {
-  var products = createFakeProducts(n);
+  var products = Generate.createProducts(n);
   var qString = 'INSERT IGNORE INTO products (description, price, imageUrl, featured, category_id, productType_id) VALUES (?, ?, ? ,? ,(SELECT id FROM categories WHERE categories.name = ?), (SELECT id FROM productTypes WHERE productTypes.name = ?))';
   for (var i = 0; i < products.length; i++) {
     await connection.query(qString, products[i], (err, res) => {
@@ -79,9 +47,6 @@ const fillProducts = async (n) => {
 
 
 module.exports = {
-  createFakeCategories,
-  createFakeProductTypes,
-  createFakeProducts,
   fillProductTypes,
   fillProducts,
   fillCategories
